实现cube坐标系转化以及计算cell之间的距离 实现A*寻路

+++ HexMap/HexGrid.cpp
@@ -14,9 +14,9 @@ void HexGrid::initGrid(int width, int height, HexCell::DirctionType type)
 {
     qDeleteAll(m_grid);
     m_grid.clear();
-    for (size_t y = 0; y < height; y++)
+    for (int y = 0; y < height; y++)
     {
-        for (size_t x = 0; x < width; x++)
+        for (int x = 0; x < width; x++)
         {
             m_grid[{x, y}] = new HexCell(this, x, y, type);
         }
